# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API",
        "description": "REST API for a restaurant inventory management system",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/stock-movements": {
            "get": {
                "tags": [
                    "Stock Movements"
                ],
                "summary": "List stock movements",
                "operationId": "56bd8b3f0ff45d75d9ee169b42b7b511",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "purchase",
                                "consumption",
                                "adjustment",
                                "waste",
                                "return"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of movements"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Stock Movements"
                ],
                "summary": "Record a new stock movement",
                "operationId": "26866531d617c2f792191d2714bf943e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "type",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "purchase",
                                            "consumption",
                                            "adjustment",
                                            "waste",
                                            "return"
                                        ]
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "minimum": 0.001
                                    },
                                    "unit_cost": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "reference_number": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "occurred_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movement recorded"
                    },
                    "422": {
                        "description": "Validation error or insufficient stock"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/stock-movements/{stock_movement}": {
            "get": {
                "tags": [
                    "Stock Movements"
                ],
                "summary": "Show a single stock movement",
                "operationId": "cf09db8d0ab090234f4120f1e3c148e2",
                "parameters": [
                    {
                        "name": "stock_movement",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movement details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/units-of-measure": {
            "get": {
                "tags": [
                    "Units of Measure"
                ],
                "summary": "List all units of measure",
                "operationId": "a75fcff66550d144de5447fe0fadad19",
                "responses": {
                    "200": {
                        "description": "Paginated list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Units of Measure"
                ],
                "summary": "Create a unit of measure (admin only)",
                "operationId": "2db0d4ba3060f2623119dd6927c82e01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "abbreviation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Kilogram"
                                    },
                                    "abbreviation": {
                                        "type": "string",
                                        "example": "kg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/units-of-measure/{unit_of_measure}": {
            "get": {
                "tags": [
                    "Units of Measure"
                ],
                "summary": "Show a unit of measure",
                "operationId": "f811fdc32c5eefc24cc27d1610329dd8",
                "parameters": [
                    {
                        "name": "unit_of_measure",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unit details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Units of Measure"
                ],
                "summary": "Update a unit of measure (admin only)",
                "operationId": "fa43c245318b620fd4027d60ebc9c1a1",
                "parameters": [
                    {
                        "name": "unit_of_measure",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "abbreviation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "abbreviation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Units of Measure"
                ],
                "summary": "Delete a unit of measure (admin only)",
                "operationId": "a8b8af6a5b8cf62e30c4efd7006447ea",
                "parameters": [
                    {
                        "name": "unit_of_measure",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}